Former-commit-id: 9c28b3ea45db1267d0c14236bb06f6171e01890d

+++ libraries/data/xmipp_polynomials.h
@@ -62,7 +62,7 @@ class PolyZernikes: public Polynomials
 {
 
 private:
-	std::vector<Matrix2D<size_t> > fMatV;
+	std::vector<Matrix2D<int> > fMatV;
 
 public:
 	//Create not really the polynomials, This function creates a set of coefficient matrix that are efficient
